+++ src/FeatureToggleSolution/FeatureToggle.Core/AlwaysOffFeatureToggle.cs
@@ -1,0 +1,11 @@
+﻿namespace FeatureToggle.Core
+
+{
+    public class AlwaysOffFeatureToggle : IFeatureToggle
+    {
+        public bool FeatureEnabled
+        {
+            get { return false; }
+        }
+    }
+}