+++ src/FeatureToggleSolution/FeatureToggle.Core/AlwaysOnFeatureToggle.cs
@@ -1,0 +1,10 @@
+﻿namespace FeatureToggle.Core
+{
+    public class AlwaysOnFeatureToggle : IFeatureToggle
+    {
+        public bool FeatureEnabled
+        {
+            get { return true; }
+        }
+    }
+}