+++ src/FeatureToggleSolution/FeatureToggle.Core/IBooleanToggleValueProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace FeatureToggle.Core
+{
+    public interface IBooleanToggleValueProvider
+    {
+        bool EvaluateBooleanToggleValue(IFeatureToggle toggle);
+    }
+}