+++ src/FeatureToggleSolution/FeatureToggle.Core/IFeatureToggle.cs
@@ -1,0 +1,7 @@
+﻿namespace FeatureToggle.Core
+{
+    public interface IFeatureToggle
+    {
+        bool FeatureEnabled {get;}
+    }
+}