+++ src/FeatureToggleSolution/FeatureToggle.Core/INowDateAndTime.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace FeatureToggle.Core
+{
+    public interface INowDateAndTime
+    {
+        DateTime Now { get;} 
+    }
+}