+++ src/FeatureToggleSolution/FeatureToggle.Core/ITimePeriodProvider.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace FeatureToggle.Core
+{
+    public interface ITimePeriodProvider
+    {
+        Tuple<DateTime, DateTime> EvaluateTimePeriod(IFeatureToggle toggle);
+    }
+}