+++ src/FeatureToggleSolution/FeatureToggle/AppSettingsProvider.cs
@@ -2,6 +2,7 @@
 using System.Configuration;
 using System.Globalization;
 using System.Linq;
+using FeatureToggle.Core;
 
 namespace JasonRoberts.FeatureToggle
 {