+++ src/FeatureToggleSolution/FeatureToggle/BooleanSqlServerProvider.cs
@@ -1,6 +1,7 @@
 ﻿using System.Configuration;
 using System.Data.SqlClient;
 using System.Linq;
+using FeatureToggle.Core;
 
 namespace JasonRoberts.FeatureToggle
 {