+++ Release-Notes.md
@@ -4,11 +4,13 @@
 
 ## Breaking Changes
 
-* 
+* Namespace changes, root namespace is now simply FeatureToggle
+* Removed Windows Phone 7 support
 
 ## Improvements
 
 * Added CLSCompliant attribute [#8](https://github.com/jason-roberts/FeatureToggle/pull/8)
+* New portable core??
 
 ##New Features